Attempt no. 2 at portable remapping of console I/O streams
Using freopen and then closing the stdio streams for the console on exit,
as in my previous attempt, causes an error during CRT shutdown. This
reprise is more involved, using and involves preserving/restoring the
original file descriptors, but is portable across VS2013 and VS2015, and
leaves the file descriptors in a correct state on exit so the CRT shuts
down cleanly.
After this change Andy's temporary change to the boot scripts can be
reverted, and the switchover to VS2015 can go ahead.