Fix issue #83 (slow/non-terminating Smalltalk syntax colouring)
Turns out this was because the workspace views are saved down with
empty whitespace character sets, which are used by the styler to establish
a start position when there is no whitespace style. Fix by treating
an empty whitespace chars setting as if it were nil, in which case the
default set of whitespace chars are used.