Replace OSVERSIONINFO/GetVersionEx with use of Version Helper APIs
GetVersionEx is deprecated. Replace explicit tests on version number
info returned by this call with calls on the Version Helper functions
now exported from the VM.
Define capability queries for most of the situations where OS version
tests existed in the class library, and use these instead to centralise
OS version checks in the SystemMetrics class. Most of the Wine fixes
are now abstracted out of the class library.