Fix compilation issue in debug config with VS2015 update 2
New error with VS2015 update 2 caused by using the runtime smaller type
check in conjunction with standard libraries. We still want to use the
check as it catches cases where data is lost by casting to a smaller
type. Apparently the STL can fire this check sometimes, hence the new
compiler error. This is not a problem for us in practice, so suppress it
by defining the pre-compiler symbol as instructed.
Note this only affects the debug build, as we only use the runtime check
in that build.