Re-enable two failing LinkButton tests and add more specific failing test
LinkButton uses the SysLink control. It looks like this has a bug whereby
it returns the length of the text plus the null terminator, whereas it
should only return the text length as other windows/controls do. I've
added a specific test for this issue, and removed the assertions failing
because of this from the two disabled tests. The specific test is added
to the known failing tests, and I'll raise an issue for that.