ProcessTest>>testCopyExceptionHandlers should not detach test runner
The purpose of this test is to verify that the exception handler stack
of a process copy (i.e. a continuation) is correctly duplicated. It does
this by creating a copy of the main process, which it terminated. The
effect of terminating a process copy will be to run any unwind blocks set
up in the origin process. If the test has been launched from the TestRunner
this would include running the #ensure: block in TestRunner>>runSuite:
which tears down the observer registration it has set up with the TestSuite.
The effect of this is that the runner stops receiving update events, and
so stops refreshing its display.