Refactor ClassCommentPlugin to introduce base for textual class plugins
Refactoring made while experimenting with adding a Class Initializer plugin
for editing a class initialization expression. Not sure about the
experiment, but the refactoring has some improvements such as fixing the
status bar display: The CHB status bar showed the first error message in the
current method and also whether the source of that method had been
changed. It showed this regardless of the card on top. Since some of the
other cards (e.g. the class comment, and class definition) support editing,
these status fields should change to reflect the status of the current plugin.