Micro perf optimizations for Smalltalk Parser.
Ideally would like to remove the need for StNumberLiteralToken to hold a
copy of the source text as there will be one of these for each number in a
method, including all the byte values in a byte array. For a method that
contains a large literal byte array, e.g. to hold an image, the overhead
of creating and storing all those strings is very large.
At present the copy of the source is needed to allow refactorings such
as inline method to preserve the original source form, otherwise qualities
such as a radix specification for an integer would be lost. The correct
fix for this is to scan numbers properly (rather than sending off to
Number>>readFrom:), and retain the extra data such as radix in a special
token types, e.g. StIntegerLiteralToken, StRadixIntegerLiteralToken,
StFloatLiteralToken, StScaledDecimalLiteralToken, etc