Remove unecessary limit on bytecode length of methods (#70)
* Remove unused DolphinIF module

* Removed unused LongPopStoreStatic bytecode and relax bytecode limit

3-byte LongPopStoreStatic instruction is not actually generated by the
compiler (and not needed, since in practice it is extremely unusual to
have a method large enough to have a literal frame with more than 256
entries). There is only one method with LongStoreStatic in the current
image, so no reason to have an optimized form with Pop.
Relaxed the bytecode limit imposed on methods with prevented the compiler
generating more than 32k of unoptimized instructions. The real limit is
on the length of jumps, so only report the "method too long" error if there
are any jumps which would be too long. This would allow large pool
initializers to be compiled without breaking down into chunks, which would
speed up compiling packages with large pool dictionaries if we changed the
file out format a bit.