Store inst vars as an Array rather than a string
This is to speed up parsing (part of work in progress on replacing the
existing Smalltalk compiler written in C++). Splitting the inst var
string is relatively expensive and the minor memory saving does not seem
the right trade off any more. The Behavior code supports both forms
interchangeably though.