Shared pool declaration order is significant but not maintained #176
Store the shared pool names as an array, and change accessors to returns
arrays rather than Sets. In order to support namespaces in future it will
probably be better to store the actual pool dictionaries rather than the
names (as this will make it easy to generate the right relative qualified
name, depending on the context), but this fixes the issue for now.