Remove redundant and subtly-wrong ListView>>selectionByIndex:ifAbsent:.
Super implementation on ListControlView already forwards it to #selection*s*ByIndex:ifAbsent: in a reasonable way.
The separate implementation has a subtle bug: Given a list with multiple selections, sending it `selectionByIndex: <one of the indices that is selected>` would leave the selection unmodified rather than selecting only the single index passed.