Fix ListView change from multi->single select not removing extra selections
There is a comment in ListView>>isMultiSelect: that says:

 "If switching to single-select from multi-select we must remove all but
  one of the selections, since the view does not do this itself."

This is true. The list view control leaves behind multiple selections if
present (which seems like a bug). Perhaps at one point Dolphin worked
around this control bug (hence the comment), but if it did the workaround
has been removed. This change reinstates or adds the selection change
described in the comment. The primary selection is retained (i.e. the one
with focus, not necessarily the first selection).