Target 10.0.14393.0 Windows SDK (#97)
Update to later Windows SDK. This is the SDK that is installed by default by VS2015 Update 3. The benefit of the change is that it is not necessary to change the default installation options for new VS installs. The downside is that if you have an older installation that you have not updated, you will need to update to U3.