Optimize Package>>allPrerequisites, #hasCyclicPrerequisites, and #hasUncommittedPrerequisites (#254)
* Optimize Package>>hasCyclicPrerequisites and #allPrerequisites.

Old implementation appeared exponential in some combination of the number of packages installed, the number of dependency relationships between them, the average number of dependencies per package, etc. Improved implementation scales linearly with the number of packages involved.

* Dramatically improve performance of Package>>hasUncommittedPrerequisites by just deferring to #allPrerequisites.

Status icons in the Package Browser, which used to lock up the image for minutes when viewing a package with many prerequisites, now add almost no lag.