Normalize line endings when filing in .st files of unknown source
Add an option to chunk source filer to normalize line endings when reading
chunks, and use this from commands that file in individual files (though
not when filing in packages, since it is slower).
Fix the fact that SequencedStream>>nextLine does not work for single-CR
line endings (like those from Squeak/Pharo).
Optimize a number of methods across the Stream hierarchy.