Delays scheduled to wake after ~49 days of uptime may fail to fire
* Fix millisecond clock wraparound issue #100 

Dolphin has always used one or other of Windows 32-bit millisecond clocks,
but these have the issue that they wraparound after ~49 days. This causes
some subtle bugs such as Delays that are scheduled before the clock
wraps, but which are expected to fire after it has wrapped, may never
actually fire.
This commit modifies the VM primitive associated with providing a
millisecond clock value to use the high resolution performance counter.
Although the resolution of this counter is much higher (on a typical
system, less than 1uS), because it is a 64-bit value it will not wrap for
many millions of days of uptime.

* High-res counter API params should really be unsigned values

QueryPerformanceCounter/Frequency seem to be wrongly declared to return
(via out parameters) signed 64-bit values, but this is nonsensical. Clearly
the h/w clock accessed by the RDTSC instruction is an unsigned value.

* Forgot to include interprt.h...