Remove semaphore parameter from primitiveSignalAtTick (#103)
There is only one Semaphore used to manage Delays, and it is stored in
the VM Pointers array, so there is no need for the primitive to take this
as an argument.

Also get rid of the OA interlocked operations - these were implemented
primarily for Windows 95, which lacked an InterlockedCompareExchange,
and also to provide a slightly more efficient InterlockedExchange. The
compiler now provides these as intrinsics, and can generate much more
efficient code.