primitiveIndirectDwordAtPut corrupts stack if can't write to memory
The ExternalAddress>>dwordAtOffset:put: primitive adjusts the stack to
early, assuming that the attempt to write to a memory location will
succeed, but if it doesn't (i.e. the memory is inaccessible), when the GP
fault is reported the Smalltalk stack is incorrect. This shows up
characteristically as the receiver being a SmallInteger, but executing
an ExternalAddress method.
There are probably other instances of this to fix.