primitiveEnableInterrupts always returns false
The underlying C++ function called by the assembler primitive is declared
as returning the C++ type "bool", which is a single byte type. The C++
compiler generates code that does not clear the upper registers of EAX -
it just sets AL. The primitive assumes it is getting a 32-bit value in EAX
so it thinks that the return value from the C++ function is always true,
which it then logically negates to generate the primitive return value
which is then always false. This results in the code in
InputState>>windowAt:put: always leaving interrupts disabled, which delays
the detection of asynchronous events until interrupts are again enabled
(e.g. by a Semaphore>>wait or signal operation).