Turn on package status icons in package browser by default
With @daniels220's perf fix for the Package prereqs tracing in PR #254,
there is no longer a performance reason not to show the package status
icons in the package browser by default. There are also a few further
related micro-optimisations in this commit to determined by profiling
the pre-requisites calculation.