Fix "MemoryManager guid is corrupted during boot for some reason ..."
This has been logged as VM issue 108, but in fact it is a problem in the
ClassBuilder: When it rebuilds a class it ensures that the VM registry
is kept up to date, however it does this by a simple name match against
the registry keys, assuming that any entry with the same name as the class
is that class. This isn't try for the "MemoryManager" entry, which is
supposed to be the instance, not the class.