Correct the visitor pattern implementation in StProgramNodeVisitor
In the visitor pattern the visitee should "accept" the visitor, and the
visitor should "visit" the visitee. The names of most of the visiting
methods in the StProgramNodeVisitor hierarchy did not follow this pattern.
This is work in progress related to #265 and should make it easier to
determine what has changed in the Refactoring Browser code.