View>>defaultId can exceed 65535, which is not valid for DM_GETDEFID
Default button Ids are manipulated using an old Windows API that limits
the button id to a 16-bit integer. The id's used in Dolphin are generated
from a monotonically increasing counter, and in a very long running
session in which many thousands of windows are opened/closed the counter
can eventually exceed 65535, which then causes a failure opening
subsequent dialogs/windows that have a default button.