Scale 16-bit identity hash value to avoid poor perf in large Sets
The 16-bit identity hash is now allocated on demand using a simple pseudo
random number generator to give reasonable distribution, but it is still
limited to 65535 values. When used as the hash for large collections this
can cause very poor performance because the objects compete for the first
64k slots, requiring a linear search through many thousands of items on
average to find the matching element. Although we cannot increase the
number of variations (since only 16-bits are available to store the
identity hash in the object memory), we can scale the value to better
distribute objects in large collections and reduce the number of probes
by orders of magnitude in the worst cases.