Allow Behavior.methodDictionary to be nil
It is reasonably commonplace for Behaviors to have no methods, especially
on the class side. For this reason the Application Deployment Wizard has
an option to fold empty method dictionaries, which is somewhat unsafe as
it will not work in applications that include the compiler and want to
add methods at runtime. By allowing the methodDictionary inst var to be
nil we no longer need this option (as there are no empty dictionaries to
fold), we save a little bit on the memory used by new classes and the time
taken to create them, and we even speed up the search for inherited methods
where there are intervening Behaviors with no methods. The cost is adding
a few nil checks, and none of these is likely to have a practical impact
on performance.