Remove most of the (significant) overhead of the isWine test
Although NTLibrary class>>isWine is not that expensive individually,
it is sprinkled among methods that are called a lot, e.g. for list view
refresh. This change makes NTLibrary a permanent library (to remove the
lazy access overhead in general, since this is a core system library)
and caches the result of the test.