Consolidating some base system changes before updating the boot image (#315)
* Add simple support for the Windows Data Protection API

Provides a simple mechanism for encrypting per-user data for persistance.

* Report struct uninitialized external call primitive failure

* XFORM metafiles class is in wrong package

* Micro-optimization of LV/TVDISPINFO notifications

Avoid creating unnecessary notification header struct.

* Remove most of the (significant) overhead of the isWine test

Although NTLibrary class>>isWine is not that expensive individually,
it is sprinkled among methods that are called a lot, e.g. for list view
refresh. This change makes NTLibrary a permanent library (to remove the
lazy access overhead in general, since this is a core system library)
and caches the result of the test.

* Add non-browsable flag to AttributeDescriptor

Non-browseable attributes are not shown in printString's or in the
published aspect inspector.

* Add Compiler class>>evaluate:for: overload

* Remove unused INTLIST theme structure