Default processing of WM_CTLCOLORXXX by parent rather than control
Windows sends the WM_CTLCOLORXXX family of messages to the parent window
of certain controls to allow the parent to set the background brush and
foreground colour. Dolphin reflects these messages to the controls
themselves to allow fine-grained colour settings, per-view. However if the
control view has no backcolor set, the message should be passed back to the
parent view for default window processing, not the control's own window
proc.