Delegate default window processing to WindowsEvent
Some WindowsEvents are forwarded by Dolphin to child windows. These should
be passed to the default window procedure of the original window to which
they were targetted, not the child. To make this less error prone, store
the original View in the event (rather than the window handle) and always
ask the event to trigger the default window processing.