Speed up slow test PushDownMethodTest>>testPushDownMethodOnNotAbstractClass
The test attempts to push down a method on Object, which is a particularly
slow case as it has to parse all methods of all subclasses. The case is
just checking that an attempt to push down a method from a non-abstract
class will fail, so any non-abstract class will do as the test subject.