Speed up SortedCollection>>add: when needs to insert before end (#339)
OrderedCollections store all elements in a contiguious block and so need
to shift ranges of elements when inserting at any position other than at
the end (or at the start when there is free space at the start). The
existing implementation did this using a loop in Smalltalk to move the
elements one by one. However there is a fast block move primitive which
can be used for this purpose, and it handles overlapping moves.