Add temp variable storage analysis based on that in Dolphin compiler
A further step in improving semanatic analysis in order to implememnt a full
Smalltalk compiler in Smalltalk. With this change the storage requirements
for each temporary variable are determined. Later this will be used to
generate the correct method header and blockcopy instructions in the
compiler backend.

All semantic analysis is currently done in the single visit of the AST.
This is adequate to determine temp usage, but performing the closure analysis
as a second pass would allow for fixing temp declarations that are at too
broad a scope by pushing them down into the most nested scope that
encompasses all accesses.