Improve the list of selectors offered when autocompleting against a receiver of known type.
Just go through all selectors in the system, asking the inferred class of the receiver if it #canUnderstand: that message. This allows classes to claim to understand methods they do not actually implement (presumably via #doesNotUnderstand:) and have those selectors appear in autocomplete lists. It also dramatically simplifies the code. (When this method was originally written, the optimization of traversing the class hierarchy for just the selectors implemented there was probably worthwhile. Now, the time taken to check all selectors is negligible.)

Also, don’t consider a variable that is known to have the value `nil` as being specifically of type UndefinedObject. Instead, treat it as being of unknown type and offer all selectors matching the prefix.