Fix TreeView/TreeListPresenter adding duplicate first level child nodes
When TreeView is configured with hasLinesAtRoot = false, it auto-expands
the first level children for newly added roots. The TLP adds a chain
of nodes up to and including a root if a new node is added by to the model
which requires intermediate nodes that are not yet in the tree. It doesn't
account for the TreeView's behaviour and re-adds the nodes that have just
been added by the TreeView under the new root node it has added. Really
the TreeView should detect and ignore the duplicate add when responding
to the add event. Since this adds a further need to look up by object in
the TreeView I decided to add a reverse map from object->handle, rather
than continuing to use a serial search across the values of the handle to
object map.
I've modified the PackagedClassSelectorTest so that it acts as a
regression test for the bug by configuring it not to show Object initially
so that the TLP has to add this when the test packaged classes are created.