Fix #407: Cursor flash in scintilla editors (#412)
Caused by the SeeingTheObjectsInside splash setting the global cursor when
it is dismissed. It should rather be clearing the global cursor it set
previously (or ideally overriding #onGetCursor:, and not setting the global
cursor at all since the global cursor mechanism is not really intended for
the purpose of setting a temporary cursor which the mouse is over a
particular area of a window). I've fixed it the cheaper way.
Also improved the printString of system cursors, as this makes it easier
to tell what is going on when such issues arise.