Add custom events storage to MemoryManager to reduce event trigger o/h
We added an event to MemoryManager that is triggered on OT overflow.
The normal event storage is in a shared dictionary with access guarded
by a mutex. This adds quite a lot of overhead to the overflow handler
when usually there are no event listeners. By providing custom event
storage in an inst var this can be avoided.