Fix bug in DiskVolumeInformation: fileSystemFlags always zero
Due to a typo in #getVolumeInformation, the fileSystemFlags inst var
is passed to the API call rather than a DWORD buffer to receiver the value.
The call doesn't error because the initial value of the inst var is nil,
which gets passed to the API as NULL, and the parameter is optional.