Implement StdioFileStream class>>on:text: for opening stream on a File
For compatibility with FileStream, provide an implementation of the
instance creation method for attaching a FileStream to an existing open
operating system File.

Also implement a much more efficient #upToEnd, because the inherited
implementation is character-by-character loop testing #atEnd before
reading each remaining character, but #atEnd is very slow on
StdioFileStreams (which also needs investigating, but is probably because
it has to flush the stream and read the file length on each call as
implemented. This is only really necessary when the buffer has been dirtied
by writing beyond the previous end of file, but the class does not track
this (unlike FileStream).