Fix Class Rename Refactor with Resources
Fix for class rename refactor not changing class references in view resource code.  I added a replace for 'Smalltalk.' , className . This now changes class references in view resources.  

It seems impure to hard code the 'Smalltalk.' prefix.  I think this notation was a starting point for potentially supporting multiple namespaces but I could not find an easy way to get a classes namespace.  Someone with a broader understanding may be able to improve the elegance of this fix.