Behavior is not abstract and should not have class and pool var behavior
Behavior represents the essential characteristics of a class from the
point of view of the execution machinery (i.e. the VM). It is not abstract
and it should be possible to create instances of it for meta-programming
purposes. It also shouldn't contain incomplete implementations of code
for accessing shared variables - this belongs in ClassDescription.