Address longstanding todo in SequenceableCollection>>copyReplaceAll:with:
SequenceableCollection>>copyReplaceAll:with: has had a todo for many years
to address its inefficient implementation that searches for the substring
occurrences to replace twice.

The todo suggested either using a stream, or building a collection of the
indices of the substrings. In exploring the use of a stream adding
additional tests I fixed the bug that write-streaming over
OrderedCollections did not work (caused by the #resize: override being
incorrect). However, although the stream implementation
is elegant, it was actually slower than the existing implementation for
many common cases, so I discarded that and went with the longer
implementation that first builds a sequence of the indices of the
occurrences of the elements to be replaced.