Fix broken version resource test caused by wide->byte string conversion
The string table loaded from the version resource is in wide string
format as this is the native format. Converting this to a byte string
causes the test to fail because some of the dolphin .exe stubs contain
a copyright symbol that is not a valid UTF8 encoding. We shouldn't need
to map the UnicodeStrings to Strings in order to be able to work with the
string table, so leave it as is.