Replace most of the Win32 ANSI API calls with Unicode equivalents
This is a fairly complete trawl through the Win32 API layers to replace
use of the ANSI functions with the Unicode versions. All of the
corresponding structures have been updated accordingly.
There is quite a lot of potential impact here, but to reduce the immediate
compatibility problems I have for the most part forced string results
from API calls back into ANSI. Once the rest of the system (in particular
Streams) can deal with UTF string encodings, the strings should probably
be converted to/from UT8 at API boundaries.
Some significant low-level Windows operations (message pump, and window
procs for Dolphin's own windows) are still in ANSI mode. This is to be
changed next, and all Windows will operate in Unicode mode. At present
only Static controls are created as Unicode, as a test case.