primitiveSize should return element count for Utf16String, not byte count
Return the number of code units in the string, not the number of bytes.
Temporarily using the string encoding for this. I will rework the instance
specification flags later to be able to free up some bits to use for this
instead (a number of the pointer/byte object flags are disjoint and so
can be overlayed).