Fix repeated walkbacks in PAI if aspect accessors fail
The PAI uses ListViews, which keep requesting the item text. If the
text block throws an error, then the ListView just keeps requesting the
text and one gets error after error. Since the PAI is inspecting a live
object using a statically calculated set of aspects, it is quite possible
for it to get out of date and need to be refreshed. It should be robust to
this.