Break dependency on icuuc.DLL, as this is only available on Windows 10 CU
Temporarily including a file from ICU4C to break the dependency on
ucuic.dll, which is only distributed with Windows 10 Creators Update.
The longer term fix is to either remove the usage of icu4c, or (more
likely) statically link the library. Unfortunately it is only included
in the Windows SDK in DLL linked form, and there does not seem to be
a nuget for a statically linked library, at least on initial examination.