Remove DolphinSureCrypto.dll and related items (#570)
* Use Windows BCrypt API for SHA1

Replaces Dolphin SHA1 use of the DolphinSureCrypto.dll with use of the standard Windows BCrypt API. Implementation of the BCrypt is partial, only for SHA1 usage, for now. The DolphinSureCryptoLib.dll is not removed from the VM yet. Will be in next PR.

* Updates to PR Use Windows BCrypt API for SHA1 #542

Implemented requested changes.

* Remove DolphinSureCrypto.dll and related

Has previously been replaced with BCrypt class using Windows API.