Default font, Trebuchet MS, has insufficient Unicode coverage (#581)
Trebuchet MS has only limited coverage of the full range of Unicode code
points. Segoe UI (the default Windows font for a few releases) is probably
the best default choice. Chose 10 point as 9 is a little small.

Addressed the issue of some dialogs having fixed size elements that don't
look good with a larger font than the default font of the original
design, by setting a particular font in those that didn't have a font
specified. Updated the font in some of those that did but which were using
the old MS Sans Serif font, or sometimes Tahoma. The dialog font is now
consistently Segoe UI 9 point.

Also addressed a small geometry issue in the calculations in TabViewXP,
causing it to lose the tails of descenders in Segoe UI 10 points - this
was hidden by Trebuchet MS 10 point as it isn't as the descenders are
shorter.