MethodHistoryBrowser won't restore deleted unmodified methods
If there is only one source entry for a method and it is deleted, the
MethodHistoryBrowser will not enable the Restore button because it thinks
the deleted version is the current version. Change the command query to
compare against the source of the method in the image (if any). This has
the additional benefit of highlighting previous versions that are the
same as the current one.