`VMTest>>testBSTRArg` is randomly very slow on AppVeyor due to XML DOM
For some reason, perhaps something to do with machine configuration,
tests that use the MSXML DOM are occasionally slow to execute on AppVeyor.
VMTest>>testBSTRArg was just using the XML DOM as a means of invoking an
external method with a bstr arg, but we can just compile one up on the fly
to avoid this.