Prevent double refresh of collection presenter on move up/down
Move up/down swaps two items in the list, causing two item change events.
The CollectionPresenter responds to any list change event (add,remove, or
update) by refreshing the entire list. This is undesirable when refreshing
the list is as expensive as it is in the Published Aspect Inspector.

Should speed up the slow test PublishedAspectInspectorTest>>testMoveDown.