Enhanced TextTileIcon that can be used on Toolbars, etc
Improvements to TextTileIcon to better support the wide range of Unicode
symbols and enable the use of TextTileIcon in all situations in which
a normal icon can be used:
- Support for transparent tile
- Text colour can be specified
- Text can be rotated in 90% increments
- Automatic sizing is more accurate
- Font style can be specified (bold, italic, etc)

There are many small changes to the GDI+ package required to support this.

ToolbarIconButtons can now successfully use TextTileIcons. Toolbars now
also support visual inheritance of button size so that the size of the
outer toolbar buttons dictates the size of all the nested toolbars.

There are further refinements to the Color hierarchy to reduce the type
testing on Color none and Color default and make Color objects other than
ARGB usable with GDI+.