Revert back to system icons for Error, Warning, and Notification
The question icon must always be loadable, without fail, so we need to use
IDI_QUESTION. The issue with 32x32 icons drawing using the old image,
rather than the current one, is I think due to an issue with the mix of
APIs being used to load icons and the over complex code that decides
whether to reload them or not. This all needs to be reworked as it also
causes only the mainstream sizes to be used, with the other sizes being
scaled down from a larger size even if the intermediate sizes are present
in the icon. This makes moving to (say) 24x24 icons for toolbars
unattractive because the carefully created 24x24 images are eschewed for
fuzzy scaledowns of the 32x32.