Ensure that bitmaps are not holding onto device contexts
The old decision to hold onto a DC in the Bitmap class provides a certain
convenience, but also has significant downsides. Bitmaps selected into a
DC can't be selected into another DC, for example, nor added to an
image list. Also it means that a GDI object is being consumed by the
Bitmap when it is not being used, which is highly undesirable since there
is a system wide limit of 10000 GDI objects, which is not very high.
Over the years workarounds for these issues have been added to
the image by explicitly freeing the DC where necessary. I'd like to remove
the cached DC, but this changes the behaviour in quite important ways and
is a breaking change, therefore I've just added some more of the workaround
and proactively free'd up the cached DC in any place that has likely
caused it to be allocated.