Improve the TreeView button hit-test bug (#599)
I think this is a bug in the underlying control not taking proper account
of the indent set with TVM_SETINDENT (modelled in Dolphin as the TreeView
`indentSeparation` property). This seems to offset the hot area for the
expando button as well as the label.

A workaround is to reduce the default indent so that the offset is less
intrusive. It seems best to let the control choose its default indent,
which on Windows 10 with 16x16 icons, is 19 pixels. The old default on
Dolphin was 24 pixels. A nice way to represent this is to use the control's
default if the `indentSeparation` property is nil. The TreeView is modified
to do this, and all the views in the base system modified to set the
property to nil.