Fix a couple of large address handling issues in the Tree and List Views
Handles/addresses returned by a Tree/ListView could be >2G when large
addresses are enabled, so handling these as signed integer values may
cause them to go negative. This causes a sign mismatch in different
code paths and prevents the TreeView from being able to map from a
handle to an object using its map when asked for display info for a node.

# Conflicts:
#	Core/Object Arts/Dolphin/MVP/Views/Common Controls/TreeView.cls