Rework largeintprim.cpp comparison prims into templated form
The LI relational ops were originally written in C++, but using a macro.
The result is code that is difficult to debug through, which is not so
using templates instead. It was hard to avoid losing perf here though as it
was difficult to get the compiler to generate code as efficient; I had to try
many approaches until settling on this.