Refactor the DocumentShell hierarchy to separate text/binary documents
The base DocumentShell should have been abstract, and as noted in a
comment in the Notepad sample "Override [streamIn:] purely
to excluded RichText and ancillaries from the deployed image: _In this case
the base class [DocumentShell] probably does too much_."

Fixes #674.

Also fix the management of the dirty indicator appended to the caption,
which does not update when expected.