Remove all message maps held in class variables
A pattern used in many view classes was to initialize a map of Windows
message ids to selectors. This pattern causes a problem with ensuring
that the image stripper does not remove the event handlers, as if the
class initializer is removed (which is generally desirable) then there are
no apparent references to the selectors in the image, and the methods will
therefore be eligible for removal. This has been addressed in the past by
flagging either the handlers or the initializer as 'most not strip', but
this is rather too easy to break by removing the special category
accidentally. It is better to rely on normal reachability by defining
the message id to handler selector map in a method that is in the call
graph.

Fixes #679