BCryptLibrary>>shaCreate fails on Windows Vista
This is a workaround for the problem of BCryptLibrary>>shaCreate raising
an invalid parameter error on Vista, because the code is taking advantage
of an improvement to the API made in Windows 7. A proper fix is arguably
not worthwhile since it would mean refactoring the code to manage the
lifetime of a buffer, and Windows Vista has now been out of support for
over 18 months.

Resolves #660