Remove forced stripping of most base package Development classes
When Dolphin was a commercial product, the license forbade shipping of
development classes in deployed applications in order to prevent the
creation of derivative development systems. This include both the contents
of specific development system packages, and a number of classes
in other packages (in particular the base Dolphin package) that were
considered part of the development system. Development classes were either
removed by the uninstallation of the base development system package
(removing all dependent packages too) or by virtue of inclusion in the
'Development' class category.
As Dolphin is now open sourced under MIT, former restrictions on uses of
the development system have been lifted, and so some time ago the forced
uninstallation of the Development System package was removed. However
removal of the Development class category remained, and would remove
classes such as category system, effectively breaking some other parts of
the development system (particularly compiling code).
This commit moves most classes out of the 'Development' class category,
and adjusts the packaging of many related methods to compensate so that
the development system continues to be effectively stripped for the
majority of applications that do not depend on the development system
as a prerequisite.