Custom draw of columns in ListView reports should not bleed to next column
The ListView control has sub-item custom draw when in report mode, but does
not reset the DC for each column. This means that settings for custom draw
in one column will persist to the next, e.g. switching one column to
draw text in italics will cause all subsequent columns to also render in
italics unless the font is changed again. This is rather awkward to work
with and means having to add custom draw blocks to columns just to reset
the canvas. The control wrapper should do the resetting instead.