Delete non-functioning ITC Gorisek Packages
STS and Omnibase do not work with Dolphin 7.1's UTF encoded strings, and
are not loaded in the build.
The HTTP client has been broken for some time, and again is not loaded.
The remainder of the packages, aside from Yaxo, are not loaded either.
All of these should be hosted externally, or in the other Contributions
area, should anyone want to fix them.

Resolves #580.