Fix poor hashing of SmallInteger keys
Hashed collections with SmallInteger keys perform poorly because the hash
of SmallInteger (the receiver value) is not well distributed across the
full hash range of 0..0x3fffffff for a typical set of integer values
encountered in practice.

Includes Integer>>isPrime, as this is useful for diagnostic purposes, especially
when trying to understand why hashed collections are performing poorly.

Fixes #391