Identity hash function must be called in a loop to avoid transient zero
The new identity hash function uses 16-bits of a 32-bit hash. Very
infrequently the 16-bit value can be zero, which is the same as the
initial value for an object that has not yet been allocated a hash, causing
a further allocation to occur the next time the identity hash is requested.
This resulted in an occasional failure of IdeaSpaceShellTests
testHistoryRemove, which could go away with code changes.
This bug was introduced by 5634a618