Display-It command now shows displayString, and Print-It the printString
Ctrl+D is bound to Display-It, Ctrl+P to Print-It.
The Display-It command uses the new Development System option to configure
the locale used for the displayString. By default this is set to the
Smalltalk locale so that, for example, numbers are displayed in Smalltalk
literal format, and Durations (e.g. from sending #timeToRun to a block)
are displayed in a human-friendly form.