Fix for breaks in SSW embedded text edit views caused by Color changes
It was possible to get away with passing nil to Canvas>>forecolor:
(or Canvas>>backcolor:) in the past, but this is now an error.
Previously the color argument was converted for passing to the underlying
Windows API call by sending it #asParameter, and the effect would have
been to pass RGB 0 (black) as the color. Therefore I've changed the code
to use Color black if nil is passed from higher layers as a foreground
colour.